refactor: extract apply agrs utils (pc)
